refactor: fix paths, make imports prettier
