+++ src/shared/ui/ErrorBoundary/index.ts
@@ -1,0 +1,1 @@
+export { ErrorBoundary } from './ui/ErrorBoundary';
