+++ src/shared/ui/ErrorBoundary/ui/index.module.css
