+++ src/shared/ui/InputSearch/index.ts
@@ -1,0 +1,1 @@
+export { InputSearch } from './ui/InputSearch';
