+++ src/shared/ui/InputSearch/ui/style.module.css
