+++ src/shared/ui/Loader/index.ts
@@ -1,0 +1,1 @@
+export { Loader } from './ui/Loader';
