+++ src/shared/ui/Loader/ui/style.module.css
